#25 S7 date and time support

+++ softplc/src/main/java/de/peteral/softplc/datatype/ByteConverter.java
@@ -31,7 +31,7 @@ public class ByteConverter implements Converter<Number> {
 	public void parseToBytes(String value, ParsedAddress address,
 			byte[] buffer, int offset) {
 
-		toBytes(Byte.parseByte(value), address, buffer, offset);
+		toBytes(Integer.parseInt(value), address, buffer, offset);
 	}
 
 }
